backend/requirements.txt
newspaper3k
pysbd
